Undo/Redo, user owned editing

+++ server/controllers/top5list-controller.js
@@ -39,9 +39,11 @@ updateTop5List = async (req, res) => {
     console.log("updateTop5List: " + JSON.stringify(body));
     let email = 0;
     console.log(req.userId);
-    await User.findById(req.userId, (err, user) => {
-        email = user.email;
-    }).catch(err => console.log(err))
+    while (email == 0) {
+        await User.findById(req.userId, (err, user) => {
+            email = user.email;
+        }).catch(err => console.log(err))
+    }
     if (!body) {
         return res.status(400).json({
             success: false,
@@ -50,10 +52,10 @@ updateTop5List = async (req, res) => {
     }
 
     Top5List.findOne({ _id: req.params.id }, (err, top5List) => {
-        console.log("top5List found: " + JSON.stringify(top5List));
+        //console.log("top5List found: " + JSON.stringify(top5List));
         console.log("Email: " + email);
         console.log("List owner: " + top5List.ownerEmail);
-        if(top5List.ownerEmail != email) {
+        if (top5List.ownerEmail != email) {
             console.log("Unauthorized edit attempt");
             return res.status(401).json({
                 err,
@@ -91,6 +93,14 @@ updateTop5List = async (req, res) => {
 }
 
 deleteTop5List = async (req, res) => {
+    let email = 0;
+    while (email == 0) {
+        await User.findById(req.userId, (err, user) => {
+            email = user.email;
+        }).catch(err => console.log(err))
+    }
+
+
     Top5List.findById({ _id: req.params.id }, (err, top5List) => {
         if (err) {
             return res.status(404).json({
@@ -98,17 +108,47 @@ deleteTop5List = async (req, res) => {
                 message: 'Top 5 List not found!',
             })
         }
+        //Check is request is from list owner
+        console.log("Email: " + email);
+        console.log("List owner: " + top5List.ownerEmail);
+        if (top5List.ownerEmail != email) {
+            console.log("Unauthorized delete attempt");
+            return res.status(401).json({
+                err,
+                message: 'Unauthorized delete attempt!',
+            })
+        }
+
         Top5List.findOneAndDelete({ _id: req.params.id }, () => {
+            console.log("deleted success");
             return res.status(200).json({ success: true, data: top5List })
         }).catch(err => console.log(err))
     })
 }
 
 getTop5ListById = async (req, res) => {
+    let email = 0;
+    while (email == 0) {
+        await User.findById(req.userId, (err, user) => {
+            email = user.email;
+        }).catch(err => console.log(err))
+    }
+
     await Top5List.findById({ _id: req.params.id }, (err, list) => {
         if (err) {
             return res.status(400).json({ success: false, error: err });
         }
+
+        console.log("Email: " + email);
+        console.log("List owner: " + list.ownerEmail);
+        if (list.ownerEmail != email) {
+            console.log("Unauthorized access attempt");
+            return res.status(401).json({
+                err,
+                message: 'Unauthorized access attempt!',
+            })
+        }
+        console.log("List access success")
         return res.status(200).json({ success: true, top5List: list })
     }).catch(err => console.log(err))
 }
@@ -127,10 +167,14 @@ getTop5Lists = async (req, res) => {
 }
 getTop5ListPairs = async (req, res) => {
     let email = 0;
-    console.log(req.userId);
-    await User.findById(req.userId, (err, user) => {
-        email = user.email;
-    }).catch(err => console.log(err))
+    console.log("ID: " + req.userId);
+    while (email == 0) {
+        await User.findById(req.userId, (err, user) => {
+            email = user.email;
+            console.log("Found: " + user.email);
+        }).catch(err => console.log(err))
+    }
+
     console.log(email);
     await Top5List.find({ ownerEmail: email }, (err, top5Lists) => {
         if (err) {
